Added type check to let* capital in resize-window
Whenever the user clicks inside of an emacs window, that counts as a
(read-key) and returns a seq instead of a number. When we bind `capital`
in the let*, we pre-suppose that char is a number, so then we try to call
+ on a number and a list. This would make lisp very angry and we were
stuck with a gross overlay.

This commit adds a type check in the definition of `capital`. If the char
entered is non-numeric, capital is defined as nil and the only other use
of it, the and below, is free to interpret that as a falsy value.